Trust all SSL certificates if necessary
