+++ ganttproject/src/biz/ganttproject/storage/cloud/GPCloudStorage.kt
@@ -28,6 +28,21 @@ import javafx.scene.control.Label
 import javafx.scene.layout.BorderPane
 import javafx.scene.layout.Pane
 import net.sourceforge.ganttproject.document.Document
+import org.apache.http.HttpHost
+import org.apache.http.auth.AuthScope
+import org.apache.http.auth.UsernamePasswordCredentials
+import org.apache.http.client.HttpClient
+import org.apache.http.client.protocol.ClientContext
+import org.apache.http.conn.scheme.Scheme
+import org.apache.http.conn.scheme.SchemeRegistry
+import org.apache.http.conn.ssl.SSLSocketFactory
+import org.apache.http.conn.ssl.TrustStrategy
+import org.apache.http.impl.auth.BasicScheme
+import org.apache.http.impl.client.BasicAuthCache
+import org.apache.http.impl.client.DefaultHttpClient
+import org.apache.http.impl.conn.PoolingClientConnectionManager
+import org.apache.http.protocol.BasicHttpContext
+import org.apache.http.protocol.HttpContext
 import org.controlsfx.control.HyperlinkLabel
 import java.util.*
 import java.util.concurrent.CompletableFuture
@@ -152,5 +167,33 @@ class HttpServerImpl : NanoHTTPD("localhost", 0) {
 
     onTokenReceived?.invoke(token, validity, userId)
     return newFixedLengthResponse("")
+  }
+}
+
+data class GPCloudHttpClient(
+    val client: HttpClient, val host: HttpHost, val context: HttpContext)
+
+object HttpClientBuilder {
+  fun buildHttpClient(): GPCloudHttpClient {
+    val httpClient = if (System.getProperty("gp.ssl.trust-all")?.toBoolean() == true) {
+      val trustAll = TrustStrategy { _, _ -> true }
+      val sslSocketFactory = SSLSocketFactory(
+          trustAll, SSLSocketFactory.ALLOW_ALL_HOSTNAME_VERIFIER)
+      val schemeRegistry = SchemeRegistry()
+      schemeRegistry.register(Scheme("https", 443, sslSocketFactory))
+      val connectionManager = PoolingClientConnectionManager(schemeRegistry)
+      DefaultHttpClient(connectionManager)
+    } else {
+      DefaultHttpClient()
+    }
+    val httpHost = HttpHost(GPCLOUD_HOST, 443, "https")
+    httpClient.credentialsProvider.setCredentials(
+        AuthScope(httpHost), UsernamePasswordCredentials(GPCloudOptions.userId.value, GPCloudOptions.authToken.value))
+    val authCache = BasicAuthCache()
+    authCache.put(httpHost, BasicScheme())
+    val context = BasicHttpContext()
+    context.setAttribute(ClientContext.AUTH_CACHE, authCache)
+
+    return GPCloudHttpClient(httpClient, httpHost, context)
   }
 }
